Add a 'recipes' subdirectory and link a file
Create beefwellington.html file within the new 'recipes' subdirectory.
Add a link in index.html to the new beeefwellington.html file.

The recipe of choice needed a webpage to itself. To proactively organize
any future recipes, a subdirectory specific for recipes was needed.
The beefwellington.html file should be accessible from the index.html
file, so a link from index.html to beefwellington.html was added.