Add recipe image, discription, ingredients, steps
An image of a finished Beef Wellington recipe was added for aesthetic
appeal and as a guideline to how the dish should look after preparing.

Added a description of the dish below the image. This gives an overview
of what will be made. Added an "Ingredients" header and unordered list
that details all food products that will be necessary for preparation.

Lastly, added a "Steps" heading and an ordered list that allows the user
to follow along as they prepare the dish.